+++ packages/filesystem/tsup.config.ts
@@ -4,7 +4,8 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   clean: true,
-  entryPoints: ["src/index.ts", "src/file-sharing.ts", "src/download.ts"],
+  external: ["@roamhq/wrtc"],
+  entryPoints: ["src/index.ts"],
   format: ["cjs", "esm"],
   dts: true,
 });