+++ packages/tunnels/tsup.config.ts
@@ -2,9 +2,16 @@
 
 import { defineConfig } from "tsup";
 
-export default defineConfig({
+export default defineConfig((options) => ({
   clean: true,
   entryPoints: ["src/index.ts", "src/openapi.ts"],
   format: ["cjs", "esm"],
-  dts: true,
-});
+  dts: options.watch
+    ? true
+    : {
+        entry: {
+          index: ".generated/types/index.d.ts",
+          openapi: ".generated/types/openapi.d.ts",
+        },
+      },
+}));